Fixed BB10 build errors from the merge. Added TARGET concept to rules. Equal to HOST, unless cross-compiling.

+++ runtimes/cpp/platforms/bb10/MoSync/src/SyscallImpl.cpp
@@ -170,12 +170,12 @@ SYSCALL(void, maWait(int timeout))
 {
 }
 
-SYSCALL(int, maTime())
+SYSCALL(longlong, maTime())
 {
 	return -1;
 }
 
-SYSCALL(int, maLocalTime())
+SYSCALL(longlong, maLocalTime())
 {
 	return -1;
 }
